Add and use VMReplacer, remove aws.UpgradeOpsMan
This replaces what was the aws command, which would have spun up a new
VM and deleted the old one in a single go.

The new command, replace-vm, is IaaS-agnostic and takes a config file,
which contains IaaS-specific configuration. It's this configuration
which is then used to figure out which IaaS-specific implementation to
use internally.

[#140407929]