Remove AWSClient interface
This makes the aws.Client interface more purposeful; previously it was
mostly just a wrapper around AWSClient, which is yet another wrapper
around the actual EC2 client.

This simplifies the approach: our aws.Client takes in an EC2 client,
which is a new interface to allow testing. The EC2 interface only has as
many methods as we actually need to use in the actual EC2 client.