Replace takes an image arg
This fixes something that became apparent to me when working on this
code before, which is that the config took the image name, but the
config is meant to be IaaS configuration, not configuration for a
particular method.

When the delete method came around I noticed that this was an issue but
I didn't fix it then.