Create new cliaas.Client
This is a new client which replaces the VMReplacer and VMDeleter of
yore.

This also moves checking completeness (not validity) of the config into
the ConfigFilePath unmarshaler, which already handles creating a new
Config object. This means the rest of the code doesn't have to worry
about checking for completeness of the config.

Why completeness, not validity? We don't know that the config is
necessarily valid; validity is determined by the IaaS. It may be
*completely filled out* but *entirely invalid*.