Remove vararg AWSClient options
Nothing is using these, and it's a namespace issue with them given
they're now in the cliaas package. cliaas.Timeout() only would have worked
for the awsClient, for example.

These things also weren't exposed in the config or on the command line
(again, they weren't used), so remove it all.